multi-agent-pipeline/pyproject.toml
[project]
name = "multi-agent-pipeline"
version = "0.1.0"
description = "An orchestrator-worker agent system with parallel execution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langsmith>=0.1.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workers"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
